feat: complete migration from JZZ to Tone.js
