+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "musischolar",
   "private": true,
-  "version": "0.0.1",
+  "version": "0.0.2",
   "type": "module",
   "scripts": {
     "dev": "vite",
