+++ src/components/Music/index.js
@@ -1,0 +1,9 @@
+import { VerovioRenderer }          from "./Verovio";
+import { verovioRenderOptions }     from "./verovioRenderOptions";
+
+export { VerovioRenderer, verovioRenderOptions as RenderOptions }
+
+import { Display } from "./Display";
+import { TonePlayer } from "./TonePlayer";
+
+export { Display, TonePlayer }
