--- src/components/Verovio/index.js
@@ -1,5 +1,0 @@
-import { Player }                   from "./MidiPlayer";
-import { VerovioRenderer }          from "./Verovio";
-import { verovioRenderOptions }     from "./verovioRenderOptions";
-
-export { Player, VerovioRenderer, verovioRenderOptions as RenderOptions }
